pyproject.toml
[tool.poetry]
name = "libraryapi-diplom"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.9"
djangorestframework = "3.14"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
pytils = "^0.4.1"
django-filter = "^24.3"
celery = "^5.4.0"
redis = "^5.0.8"
eventlet = "^0.36.1"
django-celery-beat = "^2.6.0"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
